requirements.txt
gradio==4.44.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
xgboost==2.0.3
Pillow==10.0.0
